@@ -4,11 +4,14 @@
 //! [wikipedia](https://en.wikipedia.org/w/index.php?title=Gell-Mann_matrices&oldid=988659438#Matrices)
 //! **divided by two** such that `Tr(T^a T^b) = \delta^{ab} /2 `.
 
-use std::iter::FusedIterator;
+use std::{
+    fmt::{self, Display},
+    iter::FusedIterator,
+};
 
 use nalgebra::{base::allocator::Allocator, ComplexField, DefaultAllocator, OMatrix};
 use rand::Rng;
-use rand_distr::{Distribution, Uniform};
+use rand_distr::{Bernoulli, Distribution, Uniform};
 
 use super::{
     field::Su3Adjoint, su2, utils, utils::FactorialNumber, CMatrix2, CMatrix3, Complex, Real, I,
@@ -199,6 +202,7 @@ pub const GENERATORS: [&CMatrix3; 8] = [
 #[deprecated(since = "0.2.0", note = "Please use nalgebra exp instead")]
 pub trait MatrixExp<T = Self> {
     /// Return the exponential of the matrix
+    #[must_use]
     fn exp(self) -> T;
 }
 
@@ -206,7 +210,7 @@ pub trait MatrixExp<T = Self> {
 /// It does it by first diagonalizing the matrix then exponentiate the diagonal
 /// and retransforms it back to the original basis.
 #[allow(deprecated)]
-impl<T, D> MatrixExp<OMatrix<T, D, D>> for OMatrix<T, D, D>
+impl<T, D> MatrixExp<Self> for OMatrix<T, D, D>
 where
     T: ComplexField + Copy,
     D: nalgebra::DimName + nalgebra::DimSub<nalgebra::U1>,
@@ -214,16 +218,19 @@ where
         + Allocator<T, nalgebra::DimDiff<D, nalgebra::U1>>
         + Allocator<T, D, D>
         + Allocator<T, D>,
-    OMatrix<T, D, D>: Clone,
+    Self: Clone,
 {
-    fn exp(self) -> OMatrix<T, D, D> {
+    #[inline]
+    fn exp(self) -> Self {
         let decomposition = self.schur();
         // a complex matrix is always diagonalizable
-        let eigens = decomposition.eigenvalues().unwrap();
-        let new_matrix = Self::from_diagonal(&eigens.map(|el| el.exp()));
+        let eigens = decomposition
+            .eigenvalues()
+            .expect("eignen value always exist for complex matrices");
+        let new_matrix = Self::from_diagonal(&eigens.map(ComplexField::exp));
         let (q, _) = decomposition.unpack();
         // q is always invertible
-        q.clone() * new_matrix * q.try_inverse().unwrap()
+        q.clone() * new_matrix * q.try_inverse().expect("always invertible")
     }
 }
 
@@ -276,6 +283,8 @@ where
 /// );
 /// assert_eq_matrix!(m, create_matrix_from_2_vector(v1, v2), 0.000_000_1_f64);
 /// ```
+#[inline]
+#[must_use]
 pub fn create_matrix_from_2_vector(
     v1: nalgebra::Vector3<Complex>,
     v2: nalgebra::Vector3<Complex>,
@@ -287,6 +296,8 @@ pub fn create_matrix_from_2_vector(
 }
 
 /// get an orthonormalize matrix from two vector.
+#[inline]
+#[must_use]
 fn ortho_matrix_from_2_vector(
     v1: nalgebra::Vector3<Complex>,
     v2: nalgebra::Vector3<Complex>,
@@ -299,6 +310,8 @@ fn ortho_matrix_from_2_vector(
 
 /// Try orthonormalize the given matrix.
 // TODO example
+#[inline]
+#[must_use]
 pub fn orthonormalize_matrix(matrix: &CMatrix3) -> CMatrix3 {
     let v1 = nalgebra::Vector3::from_iterator(matrix.column(0).iter().copied());
     let v2 = nalgebra::Vector3::from_iterator(matrix.column(1).iter().copied());
@@ -307,6 +320,7 @@ pub fn orthonormalize_matrix(matrix: &CMatrix3) -> CMatrix3 {
 
 /// Orthonormalize the given matrix by mutating its content.
 // TODO example
+#[inline]
 pub fn orthonormalize_matrix_mut(matrix: &mut CMatrix3) {
     *matrix = orthonormalize_matrix(matrix);
 }
@@ -322,11 +336,13 @@ pub fn orthonormalize_matrix_mut(matrix: &mut CMatrix3) {
 ///     assert_matrix_is_su_3!(random_su3(&mut rng), 9_f64 * f64::EPSILON);
 /// }
 /// ```
+#[inline]
+#[must_use]
 pub fn random_su3<Rng>(rng: &mut Rng) -> CMatrix3
 where
     Rng: rand::Rng + ?Sized,
 {
-    rand_su3_with_dis(rng, &rand::distributions::Uniform::new(-1_f64, 1_f64))
+    rand_su3_with_dis(rng, &Uniform::new(-1_f64, 1_f64))
 }
 
 /// Get a random SU3 with the given distribution.
@@ -334,6 +350,8 @@ where
 /// The given distribution can be quite opaque on the distribution of the SU(3) matrix.
 /// For a matrix Uniformly distributed among SU(3) use [`get_random_su3`].
 /// For a matrix close to unity use [`get_random_su3_close_to_unity`]
+#[inline]
+#[must_use]
 fn rand_su3_with_dis<Rng>(rng: &mut Rng, d: &impl rand_distr::Distribution<Real>) -> CMatrix3
 where
     Rng: rand::Rng + ?Sized,
@@ -350,6 +368,8 @@ where
 }
 
 /// get a random [`na::Vector3<Complex>`].
+#[inline]
+#[must_use]
 fn random_vec_3<Rng>(
     rng: &mut Rng,
     d: &impl rand_distr::Distribution<Real>,
@@ -388,6 +408,9 @@ where
 ///     f64::EPSILON * 180_f64
 /// );
 /// ```
+#[allow(clippy::missing_panics_doc)] // does not panic
+#[inline]
+#[must_use]
 pub fn random_su3_close_to_unity<R>(spread_parameter: Real, rng: &mut R) -> CMatrix3
 where
     R: rand::Rng + ?Sized,
@@ -395,7 +418,7 @@ where
     let r = get_r(su2::random_su2_close_to_unity(spread_parameter, rng));
     let s = get_s(su2::random_su2_close_to_unity(spread_parameter, rng));
     let t = get_t(su2::random_su2_close_to_unity(spread_parameter, rng));
-    let distribution = rand::distributions::Bernoulli::new(0.5_f64).unwrap();
+    let distribution = Bernoulli::new(0.5_f64).expect("always exist");
     let mut x = r * s * t;
     if distribution.sample(rng) {
         x = x.adjoint();
@@ -431,6 +454,8 @@ where
 /// );
 /// assert_eq_matrix!(get_r(m1), m2, f64::EPSILON);
 /// ```
+#[inline]
+#[must_use]
 pub fn get_r(m: CMatrix2) -> CMatrix3 {
     CMatrix3::new(
         m[(0, 0)],
@@ -475,6 +500,8 @@ pub fn get_r(m: CMatrix2) -> CMatrix3 {
 /// );
 /// assert_eq_matrix!(get_s(m1), m2, f64::EPSILON);
 /// ```
+#[inline]
+#[must_use]
 pub fn get_s(m: CMatrix2) -> CMatrix3 {
     CMatrix3::new(
         m[(0, 0)],
@@ -519,6 +546,8 @@ pub fn get_s(m: CMatrix2) -> CMatrix3 {
 /// );
 /// assert_eq_matrix!(get_t(m1), m2, f64::EPSILON);
 /// ```
+#[inline]
+#[must_use]
 pub fn get_t(m: CMatrix2) -> CMatrix3 {
     CMatrix3::new(
         ONE,
@@ -562,6 +591,8 @@ pub fn get_t(m: CMatrix2) -> CMatrix3 {
 /// );
 /// assert_eq_matrix!(get_sub_block_r(m1), m2, f64::EPSILON);
 /// ```
+#[inline]
+#[must_use]
 pub fn get_sub_block_r(m: CMatrix3) -> CMatrix2 {
     CMatrix2::new(
         m[(0, 0)],
@@ -599,6 +630,8 @@ pub fn get_sub_block_r(m: CMatrix3) -> CMatrix2 {
 /// );
 /// assert_eq_matrix!(get_sub_block_s(m1), m2, f64::EPSILON);
 /// ```
+#[inline]
+#[must_use]
 pub fn get_sub_block_s(m: CMatrix3) -> CMatrix2 {
     CMatrix2::new(
         m[(0, 0)],
@@ -636,6 +669,8 @@ pub fn get_sub_block_s(m: CMatrix3) -> CMatrix2 {
 /// );
 /// assert_eq_matrix!(get_sub_block_t(m1), m2, f64::EPSILON);
 /// ```
+#[inline]
+#[must_use]
 pub fn get_sub_block_t(m: CMatrix3) -> CMatrix2 {
     CMatrix2::new(
         m[(1, 1)],
@@ -648,24 +683,32 @@ pub fn get_sub_block_t(m: CMatrix3) -> CMatrix2 {
 
 /// Get the unormalize SU(2) sub matrix of an SU(3) matrix corresponding to the "r" sub block see
 /// [`get_sub_block_r`] and [`su2::project_to_su2_unorm`].
+#[inline]
+#[must_use]
 pub fn get_su2_r_unorm(input: CMatrix3) -> CMatrix2 {
     su2::project_to_su2_unorm(get_sub_block_r(input))
 }
 
 /// Get the unormalize SU(2) sub matrix of an SU(3) matrix corresponding to the "s" sub block see
 /// [`get_sub_block_s`] and [`su2::project_to_su2_unorm`].
+#[inline]
+#[must_use]
 pub fn get_su2_s_unorm(input: CMatrix3) -> CMatrix2 {
     su2::project_to_su2_unorm(get_sub_block_s(input))
 }
 
 /// Get the unormalize SU(2) sub matrix of an SU(3) matrix corresponding to the "t" sub block see
 /// [`get_sub_block_t`] and [`su2::project_to_su2_unorm`].
+#[inline]
+#[must_use]
 pub fn get_su2_t_unorm(input: CMatrix3) -> CMatrix2 {
     su2::project_to_su2_unorm(get_sub_block_t(input))
 }
 
 /// Get the three unormalize sub SU(2) matrix of the given SU(3) matrix, ordered `r, s, t`
 /// see [`get_sub_block_r`], [`get_sub_block_s`] and [`get_sub_block_t`]
+#[inline]
+#[must_use]
 pub fn extract_su2_unorm(m: CMatrix3) -> [CMatrix2; 3] {
     [get_su2_r_unorm(m), get_su2_s_unorm(m), get_su2_t_unorm(m)]
 }
@@ -708,21 +751,19 @@ pub fn extract_su2_unorm(m: CMatrix3) -> [CMatrix2; 3] {
 /// assert_eq!(m2, reverse(m1));
 /// assert_eq!(m1, reverse(m2));
 /// ```
+#[inline]
+#[must_use]
 pub fn reverse(input: CMatrix3) -> CMatrix3 {
-    input.map_with_location(|i, j, el| {
-        if i == j {
-            el
-        }
-        else {
-            -el
-        }
-    })
+    input.map_with_location(|i, j, el| if i == j { el } else { -el })
 }
 
 /// Return N such that `1/(N-7)!` < [`f64::EPSILON`].
 ///
 /// This number is needed for the computation of exponential matrix
 #[allow(clippy::as_conversions)] // no try into for f64
+#[allow(clippy::cast_precision_loss)]
+#[inline]
+#[must_use]
 pub fn factorial_size_for_exp() -> usize {
     let mut n: usize = 7;
     let mut factorial_value = 1;
@@ -733,7 +774,7 @@ pub fn factorial_size_for_exp() -> usize {
     n
 }
 
-/// Size of the array for FactorialStorageStatic
+/// Size of the array for [`FactorialStorageStatic`]
 const FACTORIAL_STORAGE_STAT_SIZE: usize = utils::MAX_NUMBER_FACTORIAL + 1;
 
 /// Static store for factorial number.
@@ -754,6 +795,8 @@ macro_rules! set_factorial_storage {
 impl FactorialStorageStatic {
     /// compile time evaluation of all 34 factorial numbers
     #[allow(clippy::as_conversions)] // constant function cant use try into
+    #[inline]
+    #[must_use]
     pub const fn new() -> Self {
         let mut data: [FactorialNumber; FACTORIAL_STORAGE_STAT_SIZE] =
             [1; FACTORIAL_STORAGE_STAT_SIZE];
@@ -767,11 +810,14 @@ impl FactorialStorageStatic {
     }
 
     /// access in O(1). Return None if `value` is bigger than 34.
+    #[inline]
+    #[must_use]
     pub fn try_get_factorial(&self, value: usize) -> Option<&FactorialNumber> {
         self.data.get(value)
     }
 
     /// Get an iterator over the factorial number form `0!` up to `34!`.
+    #[inline]
     pub fn iter(
         &self,
     ) -> impl Iterator<Item = &FactorialNumber> + ExactSizeIterator + FusedIterator {
@@ -779,22 +825,25 @@ impl FactorialStorageStatic {
     }
 
     /// Get the slice of factorial number.
+    #[inline]
+    #[must_use]
     pub const fn as_slice(&self) -> &[FactorialNumber; FACTORIAL_STORAGE_STAT_SIZE] {
         &self.data
     }
 }
 
 impl Default for FactorialStorageStatic {
+    #[inline]
     fn default() -> Self {
         Self::new()
     }
 }
 
-impl std::fmt::Display for FactorialStorageStatic {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+impl Display for FactorialStorageStatic {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "factorial Storage : ")?;
         for (i, n) in self.iter().enumerate() {
-            write!(f, "{}! = {}", i, n)?;
+            write!(f, "{i}! = {n}")?;
             if i < self.data.len() - 1 {
                 write!(f, ", ")?;
             }
@@ -807,12 +856,14 @@ impl<'a> IntoIterator for &'a FactorialStorageStatic {
     type IntoIter = <&'a [u128; FACTORIAL_STORAGE_STAT_SIZE] as IntoIterator>::IntoIter;
     type Item = &'a u128;
 
+    #[inline]
     fn into_iter(self) -> Self::IntoIter {
         self.data.iter()
     }
 }
 
 impl AsRef<[FactorialNumber; FACTORIAL_STORAGE_STAT_SIZE]> for FactorialStorageStatic {
+    #[inline]
     fn as_ref(&self) -> &[FactorialNumber; FACTORIAL_STORAGE_STAT_SIZE] {
         self.as_slice()
     }
@@ -834,20 +885,30 @@ const N: usize = 26;
 /// Note that the documentation above explain the algorithm for exp(X) here it is a modified version for
 /// exp(i X).
 #[inline]
+#[must_use]
 #[allow(clippy::as_conversions)] // no try into for f64
+#[allow(clippy::cast_precision_loss)]
+#[allow(clippy::missing_panics_doc)] // does not panic
 pub fn su3_exp_i(su3_adj: Su3Adjoint) -> CMatrix3 {
     // todo optimize even more using f64 to reduce the number of operation using complex that might be useless
     const N_LOOP: usize = N - 1;
-    let mut q0: Complex =
-        Complex::from(1_f64 / *FACTORIAL_STORAGE_STAT.try_get_factorial(N_LOOP).unwrap() as f64);
+    let mut q0: Complex = Complex::from(
+        1_f64
+            / *FACTORIAL_STORAGE_STAT
+                .try_get_factorial(N_LOOP)
+                .expect("should exist") as f64,
+    );
     let mut q1: Complex = Complex::from(0_f64);
     let mut q2: Complex = Complex::from(0_f64);
     let d: Complex = su3_adj.d();
     let t: Complex = su3_adj.t().into();
     for i in (0..N_LOOP).rev() {
-        let q0_n =
-            Complex::from(1_f64 / *FACTORIAL_STORAGE_STAT.try_get_factorial(i).unwrap() as f64)
-                + d * q2;
+        let q0_n = Complex::from(
+            1_f64
+                / *FACTORIAL_STORAGE_STAT
+                    .try_get_factorial(i)
+                    .expect("should exist") as f64,
+        ) + d * q2;
         let q1_n = I * (q0 - t * q2);
         let q2_n = I * q1;
 
@@ -870,7 +931,7 @@ pub fn su3_exp_i(su3_adj: Su3Adjoint) -> CMatrix3 {
 /// Note that the documentation above explain the algorithm for exp(X) here it is a modified version for
 /// exp(i X).
 ///
-/// # Panic
+/// # Panics
 /// The input matrix must be an su(3) (Lie algebra of SU(3)) matrix or approximately su(3),
 /// otherwise the function will panic in debug mod, in release the output gives unexpected values.
 /// ```should_panic
@@ -886,20 +947,34 @@ pub fn su3_exp_i(su3_adj: Su3Adjoint) -> CMatrix3 {
 /// // assert_eq_matrix!(output, (matrix* i).exp(), f64::EPSILON * 100_000_f64);
 /// ```
 #[inline]
+#[must_use]
 #[allow(clippy::as_conversions)] // no try into for f64
+#[allow(clippy::cast_precision_loss)]
 pub fn matrix_su3_exp_i(matrix: CMatrix3) -> CMatrix3 {
-    debug_assert!(is_matrix_su3_lie(&matrix, f64::EPSILON * 100_f64));
     const N_LOOP: usize = N - 1;
-    let mut q0: Complex =
-        Complex::from(1_f64 / *FACTORIAL_STORAGE_STAT.try_get_factorial(N_LOOP).unwrap() as f64);
+
+    debug_assert!(
+        is_matrix_su3_lie(&matrix, f64::EPSILON * 100_f64),
+        "the matrix si not sufficiently close to an su3 lie"
+    );
+
+    let mut q0: Complex = Complex::from(
+        1_f64
+            / *FACTORIAL_STORAGE_STAT
+                .try_get_factorial(N_LOOP)
+                .expect("should exist") as f64,
+    );
     let mut q1: Complex = Complex::from(0_f64);
     let mut q2: Complex = Complex::from(0_f64);
     let d: Complex = matrix.determinant() * I;
     let t: Complex = -Complex::from(0.5_f64) * (matrix * matrix).trace();
     for i in (0..N_LOOP).rev() {
-        let q0_n =
-            Complex::from(1_f64 / *FACTORIAL_STORAGE_STAT.try_get_factorial(i).unwrap() as f64)
-                + d * q2;
+        let q0_n = Complex::from(
+            1_f64
+                / *FACTORIAL_STORAGE_STAT
+                    .try_get_factorial(i)
+                    .expect("should exist") as f64,
+        ) + d * q2;
         let q1_n = I * (q0 - t * q2);
         let q2_n = I * q1;
 
@@ -918,19 +993,29 @@ pub fn matrix_su3_exp_i(matrix: CMatrix3) -> CMatrix3 {
 /// [OpenQCD](https://luscher.web.cern.ch/luscher/openQCD/) documentation that can be found
 /// [here](https://github.com/sa2c/OpenQCD-AVX512/blob/master/doc/su3_fcts.pdf) or by downloading a release.
 #[inline]
+#[must_use]
 #[allow(clippy::as_conversions)] // no try into for f64
+#[allow(clippy::cast_precision_loss)]
+#[allow(clippy::missing_panics_doc)] // does not panic
 pub fn su3_exp_r(su3_adj: Su3Adjoint) -> CMatrix3 {
     const N_LOOP: usize = N - 1;
-    let mut q0: Complex =
-        Complex::from(1_f64 / *FACTORIAL_STORAGE_STAT.try_get_factorial(N_LOOP).unwrap() as f64);
+    let mut q0: Complex = Complex::from(
+        1_f64
+            / *FACTORIAL_STORAGE_STAT
+                .try_get_factorial(N_LOOP)
+                .expect("always exist") as f64,
+    );
     let mut q1: Complex = Complex::from(0_f64);
     let mut q2: Complex = Complex::from(0_f64);
     let d: Complex = su3_adj.d();
     let t: Complex = su3_adj.t().into();
     for i in (0..N_LOOP).rev() {
-        let q0_n =
-            Complex::from(1_f64 / *FACTORIAL_STORAGE_STAT.try_get_factorial(i).unwrap() as f64)
-                - I * d * q2;
+        let q0_n = Complex::from(
+            1_f64
+                / *FACTORIAL_STORAGE_STAT
+                    .try_get_factorial(i)
+                    .expect("always exist") as f64,
+        ) - I * d * q2;
         let q1_n = q0 - t * q2;
         let q2_n = q1;
 
@@ -950,7 +1035,7 @@ pub fn su3_exp_r(su3_adj: Su3Adjoint) -> CMatrix3 {
 /// [OpenQCD](https://luscher.web.cern.ch/luscher/openQCD/) documentation that can be found
 /// [here](https://github.com/sa2c/OpenQCD-AVX512/blob/master/doc/su3_fcts.pdf) or by downloading a release.
 ///
-/// # Panic
+/// # Panics
 /// The input matrix must be an su(3) (Lie algebra of SU(3)) matrix or approximately su(3),
 /// otherwise the function will panic in debug mod, in release the output gives unexpected values.
 /// ```should_panic
@@ -966,20 +1051,32 @@ pub fn su3_exp_r(su3_adj: Su3Adjoint) -> CMatrix3 {
 /// // assert_eq_matrix!(output, matrix.exp(), f64::EPSILON * 100_000_f64);
 /// ```
 #[inline]
+#[must_use]
 #[allow(clippy::as_conversions)] // no try into for f64
+#[allow(clippy::cast_precision_loss)]
 pub fn matrix_su3_exp_r(matrix: CMatrix3) -> CMatrix3 {
-    debug_assert!(is_matrix_su3_lie(&matrix, f64::EPSILON * 100_f64));
     const N_LOOP: usize = N - 1;
-    let mut q0: Complex =
-        Complex::from(1_f64 / *FACTORIAL_STORAGE_STAT.try_get_factorial(N_LOOP).unwrap() as f64);
+    debug_assert!(
+        is_matrix_su3_lie(&matrix, f64::EPSILON * 100_f64),
+        "the input matrix is not on the lie su3"
+    );
+    let mut q0: Complex = Complex::from(
+        1_f64
+            / *FACTORIAL_STORAGE_STAT
+                .try_get_factorial(N_LOOP)
+                .expect("always exist") as f64,
+    );
     let mut q1: Complex = Complex::from(0_f64);
     let mut q2: Complex = Complex::from(0_f64);
     let d: Complex = matrix.determinant() * I;
     let t: Complex = -Complex::from(0.5_f64) * (matrix * matrix).trace();
     for i in (0..N_LOOP).rev() {
-        let q0_n =
-            Complex::from(1_f64 / *FACTORIAL_STORAGE_STAT.try_get_factorial(i).unwrap() as f64)
-                - I * d * q2;
+        let q0_n = Complex::from(
+            1_f64
+                / *FACTORIAL_STORAGE_STAT
+                    .try_get_factorial(i)
+                    .expect("always exist") as f64,
+        ) - I * d * q2;
         let q1_n = q0 - t * q2;
         let q2_n = q1;
 
@@ -993,18 +1090,24 @@ pub fn matrix_su3_exp_r(matrix: CMatrix3) -> CMatrix3 {
 
 /// Return wether the input matrix is SU(3) up to epsilon.
 // TODO example
+#[inline]
+#[must_use]
 pub fn is_matrix_su3(m: &CMatrix3, epsilon: f64) -> bool {
     ((m.determinant() - Complex::from(1_f64)).modulus_squared() < epsilon)
         && ((m * m.adjoint() - CMatrix3::identity()).norm() < epsilon)
 }
 
 /// Returns wether the given matrix is in the lie algebra su(3) that generates SU(3) up to epsilon.
+#[inline]
+#[must_use]
 pub fn is_matrix_su3_lie(matrix: &CMatrix3, epsilon: Real) -> bool {
     matrix.trace().modulus() < epsilon && (matrix - matrix.adjoint()).norm() < epsilon
 }
 
-#[doc(hidden)]
 /// Crate a random 3x3 Matrix
+#[doc(hidden)]
+#[inline]
+#[must_use]
 pub fn random_matrix_3<R: Rng + ?Sized>(rng: &mut R) -> CMatrix3 {
     let d = Uniform::from(-10_f64..10_f64);
     CMatrix3::from_fn(|_, _| Complex::new(d.sample(rng), d.sample(rng)))
@@ -1012,10 +1115,10 @@ pub fn random_matrix_3<R: Rng + ?Sized>(rng: &mut R) -> CMatrix3 {
 
 #[cfg(test)]
 mod test {
-    use rand::SeedableRng;
+    use rand::{rngs::StdRng, SeedableRng};
 
     use super::*;
-    use crate::{assert_eq_matrix, I};
+    use crate::{assert_eq_matrix, error::ImplementationError, I};
 
     const EPSILON: f64 = 0.000_000_001_f64;
     const SEED_RNG: u64 = 0x45_78_93_f4_4a_b0_67_f0;
@@ -1027,13 +1130,16 @@ mod test {
     }
 
     #[test]
-    fn test_factorial() {
+    fn test_factorial() -> Result<(), ImplementationError> {
         for i in 0..FACTORIAL_STORAGE_STAT_SIZE {
             assert_eq!(
-                *FACTORIAL_STORAGE_STAT.try_get_factorial(i).unwrap(),
+                *FACTORIAL_STORAGE_STAT
+                    .try_get_factorial(i)
+                    .ok_or(ImplementationError::OptionWithUnexpectedNone)?,
                 utils::factorial(i)
             );
         }
+        Ok(())
     }
 
     #[test]
@@ -1131,8 +1237,8 @@ mod test {
     #[test]
     #[allow(deprecated)]
     fn exp_matrix() {
-        let mut rng = rand::rngs::StdRng::seed_from_u64(SEED_RNG);
-        let d = rand::distributions::Uniform::from(-1_f64..1_f64);
+        let mut rng = StdRng::seed_from_u64(SEED_RNG);
+        let d = Uniform::from(-1_f64..1_f64);
         for m in &GENERATORS {
             let output_r = matrix_su3_exp_r(**m);
             assert_eq_matrix!(output_r, m.exp(), EPSILON);
@@ -1150,10 +1256,10 @@ mod test {
     }
     #[test]
     fn test_is_matrix_su3_lie() {
-        let mut rng = rand::rngs::StdRng::seed_from_u64(SEED_RNG);
-        let d = rand::distributions::Uniform::from(-1_f64..1_f64);
+        let mut rng = StdRng::seed_from_u64(SEED_RNG);
+        let d = Uniform::from(-1_f64..1_f64);
         for m in &GENERATORS {
-            assert!(is_matrix_su3_lie(*m, f64::EPSILON * 100_f64));
+            assert!(is_matrix_su3_lie(m, f64::EPSILON * 100_f64));
         }
         for _ in 0_u32..100_u32 {
             let v = Su3Adjoint::random(&mut rng, &d);
@@ -1184,7 +1290,7 @@ mod test {
 
     #[test]
     fn test_is_matrix_su3() {
-        let mut rng = rand::rngs::StdRng::seed_from_u64(SEED_RNG);
+        let mut rng = StdRng::seed_from_u64(SEED_RNG);
         let m = CMatrix3::new(
             Complex::from(0_f64),
             Complex::from(0_f64),
