@@ -199,7 +199,7 @@ fn criterion_benchmark(c: &mut Criterion) {
                     LatticeStateDefault::<4>::new_determinist(1_f64, 1_f64, *i, &mut rng).unwrap();
                 let d = rand_distr::Normal::new(0.0, 0.5_f64).unwrap();
                 let e_field = EField::new_determinist(state.lattice(), &mut rng, &d);
-                e_field.project_to_gauss(state.link_matrix(), state.lattice());
+                let _ = e_field.project_to_gauss(state.link_matrix(), state.lattice());
             })
         });
     }
