@@ -1,5 +1,6 @@
 //! Represent the fields on the lattice.
 
+use std::fmt::{self, Display};
 use std::iter::{FromIterator, FusedIterator};
 use std::{
     ops::{Add, AddAssign, Div, DivAssign, Index, IndexMut, Mul, MulAssign, Neg, Sub, SubAssign},
@@ -27,12 +28,13 @@ use super::{
 #[derive(Debug, Copy, Clone, PartialEq)]
 #[cfg_attr(feature = "serde-serialize", derive(Serialize, Deserialize))]
 pub struct Su3Adjoint {
+    /// the underling representation
     data: Vector8<Real>,
 }
 
 #[allow(clippy::len_without_is_empty)]
 impl Su3Adjoint {
-    /// create a new Su3Adjoint representation where `M = M^a T^a`, where `T` are generators given in [`su3::GENERATORS`].
+    /// create a new [`Su3Adjoint`] representation where `M = M^a T^a`, where `T` are generators given in [`su3::GENERATORS`].
     /// # Example
     /// ```
     /// use lattice_qcd_rs::field::Su3Adjoint;
@@ -40,21 +42,27 @@ impl Su3Adjoint {
     ///
     /// let su3 = Su3Adjoint::new(SVector::<f64, 8>::from_element(1_f64));
     /// ```
+    #[inline]
+    #[must_use]
     pub const fn new(data: Vector8<Real>) -> Self {
         Self { data }
     }
 
-    /// create a new Su3Adjoint representation where `M = M^a T^a`, where `T` are generators given in [`su3::GENERATORS`].
+    /// create a new [`Su3Adjoint`] representation where `M = M^a T^a`, where `T` are generators given in [`su3::GENERATORS`].
     /// # Example
     /// ```
     /// # use lattice_qcd_rs::field::Su3Adjoint;
     /// let su3 = Su3Adjoint::new_from_array([0_f64; 8]);
     /// ```
+    #[inline]
+    #[must_use]
     pub fn new_from_array(data: [Real; 8]) -> Self {
-        Su3Adjoint::new(Vector8::from(data))
+        Self::new(Vector8::from(data))
     }
 
-    /// get the data inside the Su3Adjoint.
+    /// get the data inside the [`Su3Adjoint`].
+    #[inline]
+    #[must_use]
     pub const fn data(&self) -> &Vector8<Real> {
         &self.data
     }
@@ -69,6 +77,8 @@ impl Su3Adjoint {
     /// let max = adj.as_vector().max();
     /// let norm = adj.as_ref().norm();
     /// ```
+    #[inline]
+    #[must_use]
     pub const fn as_vector(&self) -> &Vector8<Real> {
         self.data()
     }
@@ -90,6 +100,8 @@ impl Su3Adjoint {
     ///
     /// assert_eq!(adj.as_vector(), &v);
     /// ```
+    #[inline]
+    #[must_use]
     pub fn as_vector_mut(&mut self) -> &mut Vector8<Real> {
         self.data_mut()
     }
@@ -103,6 +115,8 @@ impl Su3Adjoint {
     /// assert_eq!(su3.to_matrix(), *lattice_qcd_rs::su3::GENERATORS[0]);
     /// ```
     // TODO self non consumed ?? passé en &self ? TODO bench
+    #[inline]
+    #[must_use]
     pub fn to_matrix(self) -> Matrix3<nalgebra::Complex<Real>> {
         self.data
             .into_iter()
@@ -120,6 +134,8 @@ impl Su3Adjoint {
     /// let su3 = Su3Adjoint::new_from_array([1_f64, 0_f64, 0_f64, 0_f64, 0_f64, 0_f64, 0_f64, 0_f64]);
     /// assert_eq!(su3.to_su3().determinant(), nalgebra::Complex::from(1_f64));
     /// ```
+    #[inline]
+    #[must_use]
     pub fn to_su3(self) -> Matrix3<nalgebra::Complex<Real>> {
         // NOTE: should it consume ? the user can manually clone and there is use because
         // where the value is not necessary anymore.
@@ -128,6 +144,8 @@ impl Su3Adjoint {
 
     /// Return exp( T^a v^a) where v is self.
     /// Note that the function consume self.
+    #[inline]
+    #[must_use]
     pub fn exp(self) -> Matrix3<nalgebra::Complex<Real>> {
         su3::su3_exp_r(self)
     }
@@ -142,6 +160,8 @@ impl Su3Adjoint {
     /// let distribution = rand::distributions::Uniform::from(-1_f64..1_f64);
     /// let su3 = Su3Adjoint::random(&mut rng, &distribution);
     /// ```
+    #[inline]
+    #[must_use]
     pub fn random<Rng>(rng: &mut Rng, d: &impl rand_distr::Distribution<Real>) -> Self
     where
         Rng: rand::Rng + ?Sized,
@@ -161,13 +181,14 @@ impl Su3Adjoint {
     /// (m.to_matrix() * m.to_matrix()).trace().real();
     /// ```
     #[inline]
+    #[must_use]
     pub fn trace_squared(&self) -> Real {
         // TODO investigate.
         self.data.iter().map(|el| el * el).sum::<Real>() / 2_f64
     }
 
     /// Return the t coeff `t = - 1/2 * Tr(X^2)`.
-    /// If you are looking for the trace square use [Self::trace_squared] instead.
+    /// If you are looking for the trace square use [`Self::trace_squared`] instead.
     ///
     /// It is used for [`su3::su3_exp_i`].
     ///
@@ -182,6 +203,7 @@ impl Su3Adjoint {
     /// );
     /// ```
     #[inline]
+    #[must_use]
     pub fn t(&self) -> Real {
         -0.5_f64 * self.trace_squared()
     }
@@ -200,6 +222,7 @@ impl Su3Adjoint {
     /// );
     /// ```
     #[inline]
+    #[must_use]
     pub fn d(&self) -> nalgebra::Complex<Real> {
         self.to_matrix().determinant() * I
     }
@@ -210,7 +233,8 @@ impl Su3Adjoint {
     /// # let su3 = Su3Adjoint::new(nalgebra::SVector::<f64, 8>::zeros());
     /// assert_eq!(su3.len(), 8);
     /// ```
-    #[allow(clippy::unused_self)]
+    #[inline]
+    #[must_use]
     pub fn len(&self) -> usize {
         self.data.len()
         //8
@@ -222,6 +246,8 @@ impl Su3Adjoint {
     /// let su3 = Su3Adjoint::new(nalgebra::SVector::<f64, 8>::zeros());
     /// assert_eq!(Su3Adjoint::len_const(), su3.len());
     /// ```
+    #[inline]
+    #[must_use]
     pub const fn len_const() -> usize {
         8
     }
@@ -234,6 +260,7 @@ impl Su3Adjoint {
     /// # let adj = Su3Adjoint::default();
     /// let sum_abs = adj.iter().map(|el| el.abs()).sum::<f64>();
     /// ```
+    #[inline]
     pub fn iter(&self) -> impl Iterator<Item = &Real> + ExactSizeIterator + FusedIterator {
         self.data.iter()
     }
@@ -246,6 +273,7 @@ impl Su3Adjoint {
     /// # let mut adj = Su3Adjoint::default();
     /// adj.iter_mut().for_each(|el| *el = *el / 2_f64);
     /// ```
+    #[inline]
     pub fn iter_mut(
         &mut self,
     ) -> impl Iterator<Item = &mut Real> + ExactSizeIterator + FusedIterator {
@@ -253,18 +281,22 @@ impl Su3Adjoint {
     }
 
     /// Get a mutable reference over the data.
+    #[inline]
+    #[must_use]
     pub fn data_mut(&mut self) -> &mut Vector8<Real> {
         &mut self.data
     }
 }
 
 impl AsRef<Vector8<f64>> for Su3Adjoint {
+    #[inline]
     fn as_ref(&self) -> &Vector8<f64> {
         self.as_vector()
     }
 }
 
 impl AsMut<Vector8<f64>> for Su3Adjoint {
+    #[inline]
     fn as_mut(&mut self) -> &mut Vector8<f64> {
         self.as_vector_mut()
     }
@@ -274,6 +306,7 @@ impl<'a> IntoIterator for &'a Su3Adjoint {
     type IntoIter = <&'a Vector8<Real> as IntoIterator>::IntoIter;
     type Item = &'a Real;
 
+    #[inline]
     fn into_iter(self) -> Self::IntoIter {
         self.data.iter()
     }
@@ -283,29 +316,33 @@ impl<'a> IntoIterator for &'a mut Su3Adjoint {
     type IntoIter = <&'a mut Vector8<Real> as IntoIterator>::IntoIter;
     type Item = &'a mut Real;
 
+    #[inline]
     fn into_iter(self) -> Self::IntoIter {
         self.data.iter_mut()
     }
 }
 
 impl AddAssign for Su3Adjoint {
+    #[inline]
     fn add_assign(&mut self, other: Self) {
         self.data += other.data();
     }
 }
 
-impl Add<Su3Adjoint> for Su3Adjoint {
+impl Add<Self> for Su3Adjoint {
     type Output = Self;
 
+    #[inline]
     fn add(mut self, rhs: Self) -> Self::Output {
         self += rhs;
         self
     }
 }
 
-impl Add<&Su3Adjoint> for Su3Adjoint {
+impl Add<&Self> for Su3Adjoint {
     type Output = Self;
 
+    #[inline]
     fn add(self, rhs: &Self) -> Self::Output {
         self + *rhs
     }
@@ -314,6 +351,7 @@ impl Add<&Su3Adjoint> for Su3Adjoint {
 impl Add<Su3Adjoint> for &Su3Adjoint {
     type Output = Su3Adjoint;
 
+    #[inline]
     fn add(self, rhs: Su3Adjoint) -> Self::Output {
         rhs + self
     }
@@ -322,12 +360,14 @@ impl Add<Su3Adjoint> for &Su3Adjoint {
 impl Add<&Su3Adjoint> for &Su3Adjoint {
     type Output = Su3Adjoint;
 
+    #[inline]
     fn add(self, rhs: &Su3Adjoint) -> Self::Output {
         self + *rhs
     }
 }
 
 impl MulAssign<f64> for Su3Adjoint {
+    #[inline]
     fn mul_assign(&mut self, rhs: f64) {
         self.data *= rhs;
     }
@@ -336,6 +376,7 @@ impl MulAssign<f64> for Su3Adjoint {
 impl Mul<Real> for Su3Adjoint {
     type Output = Self;
 
+    #[inline]
     fn mul(mut self, rhs: Real) -> Self::Output {
         self *= rhs;
         self
@@ -345,6 +386,7 @@ impl Mul<Real> for Su3Adjoint {
 impl Mul<&Real> for Su3Adjoint {
     type Output = Self;
 
+    #[inline]
     fn mul(self, rhs: &Real) -> Self::Output {
         self * (*rhs)
     }
@@ -353,6 +395,7 @@ impl Mul<&Real> for Su3Adjoint {
 impl Mul<Real> for &Su3Adjoint {
     type Output = Su3Adjoint;
 
+    #[inline]
     fn mul(self, rhs: Real) -> Self::Output {
         *self * rhs
     }
@@ -361,6 +404,7 @@ impl Mul<Real> for &Su3Adjoint {
 impl Mul<&Real> for &Su3Adjoint {
     type Output = Su3Adjoint;
 
+    #[inline]
     fn mul(self, rhs: &Real) -> Self::Output {
         *self * rhs
     }
@@ -369,6 +413,7 @@ impl Mul<&Real> for &Su3Adjoint {
 impl Mul<Su3Adjoint> for Real {
     type Output = Su3Adjoint;
 
+    #[inline]
     fn mul(self, rhs: Su3Adjoint) -> Self::Output {
         rhs * self
     }
@@ -377,6 +422,7 @@ impl Mul<Su3Adjoint> for Real {
 impl Mul<&Su3Adjoint> for Real {
     type Output = Su3Adjoint;
 
+    #[inline]
     fn mul(self, rhs: &Su3Adjoint) -> Self::Output {
         rhs * self
     }
@@ -385,6 +431,7 @@ impl Mul<&Su3Adjoint> for Real {
 impl Mul<Su3Adjoint> for &Real {
     type Output = Su3Adjoint;
 
+    #[inline]
     fn mul(self, rhs: Su3Adjoint) -> Self::Output {
         rhs * self
     }
@@ -393,18 +440,21 @@ impl Mul<Su3Adjoint> for &Real {
 impl Mul<&Su3Adjoint> for &Real {
     type Output = Su3Adjoint;
 
+    #[inline]
     fn mul(self, rhs: &Su3Adjoint) -> Self::Output {
         rhs * self
     }
 }
 
 impl DivAssign<f64> for Su3Adjoint {
+    #[inline]
     fn div_assign(&mut self, rhs: f64) {
         self.data /= rhs;
     }
 }
 
 impl DivAssign<&f64> for Su3Adjoint {
+    #[inline]
     fn div_assign(&mut self, rhs: &f64) {
         self.data /= *rhs;
     }
@@ -413,6 +463,7 @@ impl DivAssign<&f64> for Su3Adjoint {
 impl Div<Real> for Su3Adjoint {
     type Output = Self;
 
+    #[inline]
     fn div(mut self, rhs: Real) -> Self::Output {
         self /= rhs;
         self
@@ -422,6 +473,7 @@ impl Div<Real> for Su3Adjoint {
 impl Div<&Real> for Su3Adjoint {
     type Output = Self;
 
+    #[inline]
     fn div(self, rhs: &Real) -> Self::Output {
         self / (*rhs)
     }
@@ -430,6 +482,7 @@ impl Div<&Real> for Su3Adjoint {
 impl Div<Real> for &Su3Adjoint {
     type Output = Su3Adjoint;
 
+    #[inline]
     fn div(self, rhs: Real) -> Self::Output {
         *self / rhs
     }
@@ -438,29 +491,33 @@ impl Div<Real> for &Su3Adjoint {
 impl Div<&Real> for &Su3Adjoint {
     type Output = Su3Adjoint;
 
+    #[inline]
     fn div(self, rhs: &Real) -> Self::Output {
         *self / rhs
     }
 }
 
 impl SubAssign for Su3Adjoint {
+    #[inline]
     fn sub_assign(&mut self, other: Self) {
         self.data -= other.data();
     }
 }
 
-impl Sub<Su3Adjoint> for Su3Adjoint {
+impl Sub<Self> for Su3Adjoint {
     type Output = Self;
 
+    #[inline]
     fn sub(mut self, rhs: Self) -> Self::Output {
         self -= rhs;
         self
     }
 }
 
-impl Sub<&Su3Adjoint> for Su3Adjoint {
+impl Sub<&Self> for Su3Adjoint {
     type Output = Self;
 
+    #[inline]
     fn sub(self, rhs: &Self) -> Self::Output {
         self - *rhs
     }
@@ -469,6 +526,7 @@ impl Sub<&Su3Adjoint> for Su3Adjoint {
 impl Sub<Su3Adjoint> for &Su3Adjoint {
     type Output = Su3Adjoint;
 
+    #[inline]
     fn sub(self, rhs: Su3Adjoint) -> Self::Output {
         rhs - self
     }
@@ -477,6 +535,7 @@ impl Sub<Su3Adjoint> for &Su3Adjoint {
 impl Sub<&Su3Adjoint> for &Su3Adjoint {
     type Output = Su3Adjoint;
 
+    #[inline]
     fn sub(self, rhs: &Su3Adjoint) -> Self::Output {
         *self - rhs
     }
@@ -485,14 +544,16 @@ impl Sub<&Su3Adjoint> for &Su3Adjoint {
 impl Neg for Su3Adjoint {
     type Output = Self;
 
+    #[inline]
     fn neg(self) -> Self::Output {
-        Su3Adjoint::new(-self.data)
+        Self::new(-self.data)
     }
 }
 
 impl Neg for &Su3Adjoint {
     type Output = Su3Adjoint;
 
+    #[inline]
     fn neg(self) -> Self::Output {
         Su3Adjoint::new(-self.data())
     }
@@ -510,13 +571,15 @@ impl Default for Su3Adjoint {
     ///     Su3Adjoint::new_from_array([0_f64; 8])
     /// );
     /// ```
+    #[inline]
     fn default() -> Self {
-        Su3Adjoint::new(Vector8::from_element(0_f64))
+        Self::new(Vector8::from_element(0_f64))
     }
 }
 
-impl std::fmt::Display for Su3Adjoint {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+impl Display for Su3Adjoint {
+    #[inline]
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "{}", self.to_matrix())
     }
 }
@@ -533,6 +596,7 @@ impl Index<usize> for Su3Adjoint {
     /// let su3 = Su3Adjoint::new_from_array([0_f64; 8]);
     /// let _ = su3[8];
     /// ```
+    #[inline]
     fn index(&self, pos: usize) -> &Self::Output {
         &self.data[pos]
     }
@@ -548,32 +612,37 @@ impl IndexMut<usize> for Su3Adjoint {
     /// let mut su3 = Su3Adjoint::new_from_array([0_f64; 8]);
     /// su3[8] += 1_f64;
     /// ```
+    #[inline]
     fn index_mut(&mut self, pos: usize) -> &mut Self::Output {
         &mut self.data[pos]
     }
 }
 
 impl From<Vector8<Real>> for Su3Adjoint {
+    #[inline]
     fn from(v: Vector8<Real>) -> Self {
-        Su3Adjoint::new(v)
+        Self::new(v)
     }
 }
 
 impl From<Su3Adjoint> for Vector8<Real> {
+    #[inline]
     fn from(v: Su3Adjoint) -> Self {
         v.data
     }
 }
 
 impl From<&Su3Adjoint> for Vector8<Real> {
+    #[inline]
     fn from(v: &Su3Adjoint) -> Self {
         v.data
     }
 }
 
 impl From<[Real; 8]> for Su3Adjoint {
+    #[inline]
     fn from(v: [Real; 8]) -> Self {
-        Su3Adjoint::new_from_array(v)
+        Self::new_from_array(v)
     }
 }
 
@@ -587,36 +656,50 @@ pub struct LinkMatrix {
 
 impl LinkMatrix {
     /// Create a new link matrix field from preexisting data.
+    #[inline]
+    #[must_use]
     pub const fn new(data: Vec<Matrix3<nalgebra::Complex<Real>>>) -> Self {
         Self { data }
     }
 
     /// Get the raw data.
+    #[inline]
+    #[must_use]
     pub const fn data(&self) -> &Vec<Matrix3<nalgebra::Complex<Real>>> {
         &self.data
     }
 
     /// Get a mutable reference to the data
+    #[inline]
+    #[must_use]
     pub fn data_mut(&mut self) -> &mut Vec<Matrix3<nalgebra::Complex<Real>>> {
         &mut self.data
     }
 
-    /// Get the link_matrix as a [`Vec`].
+    /// Get the `link_matrix` as a [`Vec`].
+    #[inline]
+    #[must_use]
     pub const fn as_vec(&self) -> &Vec<Matrix3<nalgebra::Complex<Real>>> {
         self.data()
     }
 
-    /// Get the link_matrix as a mutable [`Vec`].
+    /// Get the `link_matrix` as a mutable [`Vec`].
+    #[inline]
+    #[must_use]
     pub fn as_vec_mut(&mut self) -> &mut Vec<Matrix3<nalgebra::Complex<Real>>> {
         self.data_mut()
     }
 
-    /// Get the link_matrix as a slice.
+    /// Get the `link_matrix` as a slice.
+    #[inline]
+    #[must_use]
     pub fn as_slice(&self) -> &[Matrix3<nalgebra::Complex<Real>>] {
         self.data()
     }
 
-    /// Get the link_matrix as a mut ref to a slice
+    /// Get the `link_matrix` as a mut ref to a slice
+    #[inline]
+    #[must_use]
     pub fn as_slice_mut(&mut self) -> &mut [Matrix3<nalgebra::Complex<Real>>] {
         &mut self.data
     }
@@ -643,6 +726,8 @@ impl LinkMatrix {
     /// # Ok(())
     /// # }
     /// ```
+    #[inline]
+    #[must_use]
     pub fn new_determinist<Rng: rand::Rng + ?Sized, const D: usize>(
         l: &LatticeCyclic<D>,
         rng: &mut Rng,
@@ -677,16 +762,16 @@ impl LinkMatrix {
     ///
     /// # Errors
     /// Returns [`ThreadError::ThreadNumberIncorrect`] if `number_of_thread` is 0.
+    #[inline]
     pub fn new_random_threaded<const D: usize>(
         l: &LatticeCyclic<D>,
         number_of_thread: usize,
     ) -> Result<Self, ThreadError> {
         if number_of_thread == 0 {
             return Err(ThreadError::ThreadNumberIncorrect);
-        }
-        else if number_of_thread == 1 {
+        } else if number_of_thread == 1 {
             let mut rng = rand::thread_rng();
-            return Ok(LinkMatrix::new_determinist(l, &mut rng));
+            return Ok(Self::new_determinist(l, &mut rng));
         }
         let data = run_pool_parallel_vec_with_initializations_mutable(
             l.get_links(),
@@ -715,6 +800,8 @@ impl LinkMatrix {
     /// # Ok(())
     /// # }
     /// ```
+    #[inline]
+    #[must_use]
     pub fn new_cold<const D: usize>(l: &LatticeCyclic<D>) -> Self {
         Self {
             data: vec![CMatrix3::identity(); l.number_of_canonical_links_space()],
@@ -723,6 +810,8 @@ impl LinkMatrix {
 
     /// get the link matrix associated to given link using the notation
     /// $`U_{-i}(x) = U^\dagger_{i}(x-i)`$
+    #[inline]
+    #[must_use]
     pub fn matrix<const D: usize>(
         &self,
         link: &LatticeLink<D>,
@@ -733,13 +822,15 @@ impl LinkMatrix {
         if link.is_dir_negative() {
             // that means the the link was in the negative direction
             Some(matrix.adjoint())
-        }
-        else {
+        } else {
             Some(*matrix)
         }
     }
 
     /// Get $`S_ij(x) = U_j(x) U_i(x+j) U^\dagger_j(x+i)`$.
+    #[allow(clippy::similar_names)]
+    #[inline]
+    #[must_use]
     pub fn sij<const D: usize>(
         &self,
         point: &LatticePoint<D>,
@@ -758,6 +849,8 @@ impl LinkMatrix {
     }
 
     /// Get the plaquette $`P_{ij}(x) = U_i(x) S^\dagger_ij(x)`$.
+    #[inline]
+    #[must_use]
     pub fn pij<const D: usize>(
         &self,
         point: &LatticePoint<D>,
@@ -771,7 +864,8 @@ impl LinkMatrix {
     }
 
     /// Take the average of the trace of all plaquettes
-    #[allow(clippy::as_conversions)] // no try into for f64
+    #[inline]
+    #[must_use]
     pub fn average_trace_plaquette<const D: usize>(
         &self,
         lattice: &LatticeCyclic<D>,
@@ -799,11 +893,16 @@ impl LinkMatrix {
             })
             .sum::<Option<nalgebra::Complex<Real>>>()?;
         let number_of_directions = (D * (D - 1)) / 2;
+
+        #[allow(clippy::as_conversions)] // no try into for f64
+        #[allow(clippy::cast_precision_loss)]
         let number_of_plaquette = (lattice.number_of_points() * number_of_directions) as f64;
         Some(sum / number_of_plaquette)
     }
 
-    /// Get the clover, used for F_mu_nu tensor
+    /// Get the clover, used for `F_mu_nu` tensor
+    #[inline]
+    #[must_use]
     pub fn clover<const D: usize>(
         &self,
         point: &LatticePoint<D>,
@@ -822,6 +921,8 @@ impl LinkMatrix {
     /// Get the `F^{ij}` tensor using the clover appropriation. The direction should be set to positive.
     /// See <https://arxiv.org/abs/1512.02374>.
     // TODO negative directions
+    #[inline]
+    #[must_use]
     pub fn f_mu_nu<const D: usize>(
         &self,
         point: &LatticePoint<D>,
@@ -835,6 +936,8 @@ impl LinkMatrix {
     }
 
     /// Get the chromomagnetic field at a given point.
+    #[inline]
+    #[must_use]
     pub fn magnetic_field_vec<const D: usize>(
         &self,
         point: &LatticePoint<D>,
@@ -848,6 +951,8 @@ impl LinkMatrix {
     }
 
     /// Get the chromomagnetic field at a given point alongside a given direction.
+    #[inline]
+    #[must_use]
     pub fn magnetic_field<const D: usize>(
         &self,
         point: &LatticePoint<D>,
@@ -873,6 +978,8 @@ impl LinkMatrix {
     }
 
     /// Get the chromomagnetic field at a given point alongside a given direction given by lattice link.
+    #[inline]
+    #[must_use]
     pub fn magnetic_field_link<const D: usize>(
         &self,
         link: &LatticeLink<D>,
@@ -882,11 +989,15 @@ impl LinkMatrix {
     }
 
     /// Return the number of elements.
+    #[inline]
+    #[must_use]
     pub fn len(&self) -> usize {
         self.data.len()
     }
 
     /// Returns wether the there is no data.
+    #[inline]
+    #[must_use]
     pub fn is_empty(&self) -> bool {
         self.data.is_empty()
     }
@@ -894,6 +1005,7 @@ impl LinkMatrix {
     /// Correct the numerical drift, reprojecting all the matrices to SU(3).
     ///
     /// You can look at the example of [`super::simulation::LatticeStateDefault::normalize_link_matrices`]
+    #[inline]
     pub fn normalize(&mut self) {
         self.data.par_iter_mut().for_each(|el| {
             su3::orthonormalize_matrix_mut(el);
@@ -901,11 +1013,13 @@ impl LinkMatrix {
     }
 
     /// Iter on the data.
+    #[inline]
     pub fn iter(&self) -> impl Iterator<Item = &CMatrix3> + ExactSizeIterator + FusedIterator {
         self.data.iter()
     }
 
     /// Iter mutably on the data.
+    #[inline]
     pub fn iter_mut(
         &mut self,
     ) -> impl Iterator<Item = &mut CMatrix3> + ExactSizeIterator + FusedIterator {
@@ -914,24 +1028,28 @@ impl LinkMatrix {
 }
 
 impl AsRef<Vec<CMatrix3>> for LinkMatrix {
+    #[inline]
     fn as_ref(&self) -> &Vec<CMatrix3> {
         self.as_vec()
     }
 }
 
 impl AsMut<Vec<CMatrix3>> for LinkMatrix {
+    #[inline]
     fn as_mut(&mut self) -> &mut Vec<CMatrix3> {
         self.as_vec_mut()
     }
 }
 
 impl AsRef<[CMatrix3]> for LinkMatrix {
+    #[inline]
     fn as_ref(&self) -> &[CMatrix3] {
         self.as_slice()
     }
 }
 
 impl AsMut<[CMatrix3]> for LinkMatrix {
+    #[inline]
     fn as_mut(&mut self) -> &mut [CMatrix3] {
         self.as_slice_mut()
     }
@@ -941,6 +1059,7 @@ impl<'a> IntoIterator for &'a LinkMatrix {
     type IntoIter = <&'a Vec<CMatrix3> as IntoIterator>::IntoIter;
     type Item = &'a CMatrix3;
 
+    #[inline]
     fn into_iter(self) -> Self::IntoIter {
         self.data.iter()
     }
@@ -950,6 +1069,7 @@ impl<'a> IntoIterator for &'a mut LinkMatrix {
     type IntoIter = <&'a mut Vec<CMatrix3> as IntoIterator>::IntoIter;
     type Item = &'a mut CMatrix3;
 
+    #[inline]
     fn into_iter(self) -> Self::IntoIter {
         self.data.iter_mut()
     }
@@ -958,12 +1078,14 @@ impl<'a> IntoIterator for &'a mut LinkMatrix {
 impl Index<usize> for LinkMatrix {
     type Output = CMatrix3;
 
+    #[inline]
     fn index(&self, pos: usize) -> &Self::Output {
         &self.data[pos]
     }
 }
 
 impl IndexMut<usize> for LinkMatrix {
+    #[inline]
     fn index_mut(&mut self, pos: usize) -> &mut Self::Output {
         &mut self.data[pos]
     }
@@ -973,6 +1095,7 @@ impl<A> FromIterator<A> for LinkMatrix
 where
     Vec<CMatrix3>: FromIterator<A>,
 {
+    #[inline]
     fn from_iter<T>(iter: T) -> Self
     where
         T: IntoIterator<Item = A>,
@@ -986,6 +1109,7 @@ where
     Vec<CMatrix3>: FromParallelIterator<A>,
     A: Send,
 {
+    #[inline]
     fn from_par_iter<I>(par_iter: I) -> Self
     where
         I: IntoParallelIterator<Item = A>,
@@ -999,6 +1123,7 @@ where
     Vec<CMatrix3>: ParallelExtend<T>,
     T: Send,
 {
+    #[inline]
     fn par_extend<I>(&mut self, par_iter: I)
     where
         I: IntoParallelIterator<Item = T>,
@@ -1011,6 +1136,7 @@ impl<A> Extend<A> for LinkMatrix
 where
     Vec<CMatrix3>: Extend<A>,
 {
+    #[inline]
     fn extend<T>(&mut self, iter: T)
     where
         T: IntoIterator<Item = A>,
@@ -1028,36 +1154,50 @@ pub struct EField<const D: usize> {
 
 impl<const D: usize> EField<D> {
     /// Create a new "Electrical" field.
+    #[inline]
+    #[must_use]
     pub fn new(data: Vec<SVector<Su3Adjoint, D>>) -> Self {
         Self { data }
     }
 
     /// Get the raw data.
+    #[inline]
+    #[must_use]
     pub const fn data(&self) -> &Vec<SVector<Su3Adjoint, D>> {
         &self.data
     }
 
     /// Get a mut ref to the data data.
+    #[inline]
+    #[must_use]
     pub fn data_mut(&mut self) -> &mut Vec<SVector<Su3Adjoint, D>> {
         &mut self.data
     }
 
-    /// Get the e_field as a Vec of Vector of [`Su3Adjoint`]
+    /// Get the `e_field` as a Vec of Vector of [`Su3Adjoint`]
+    #[inline]
+    #[must_use]
     pub const fn as_vec(&self) -> &Vec<SVector<Su3Adjoint, D>> {
         self.data()
     }
 
-    /// Get the e_field as a slice of Vector of [`Su3Adjoint`]
+    /// Get the `e_field` as a slice of Vector of [`Su3Adjoint`]
+    #[inline]
+    #[must_use]
     pub fn as_slice(&self) -> &[SVector<Su3Adjoint, D>] {
         &self.data
     }
 
-    /// Get the e_field as mut ref to slice of Vector of [`Su3Adjoint`]
+    /// Get the `e_field` as mut ref to slice of Vector of [`Su3Adjoint`]
+    #[inline]
+    #[must_use]
     pub fn as_slice_mut(&mut self) -> &mut [SVector<Su3Adjoint, D>] {
         &mut self.data
     }
 
     /// Return the number of elements.
+    #[inline]
+    #[must_use]
     pub fn len(&self) -> usize {
         self.data.len()
     }
@@ -1083,6 +1223,8 @@ impl<const D: usize> EField<D> {
     /// # Ok(())
     /// # }
     /// ```
+    #[inline]
+    #[must_use]
     pub fn new_determinist<Rng: rand::Rng + ?Sized>(
         l: &LatticeCyclic<D>,
         rng: &mut Rng,
@@ -1114,9 +1256,11 @@ impl<const D: usize> EField<D> {
     /// # Ok(())
     /// # }
     /// ```
+    #[inline]
+    #[must_use]
     pub fn new_random(l: &LatticeCyclic<D>, d: &impl rand_distr::Distribution<Real>) -> Self {
         let mut rng = rand::thread_rng();
-        EField::new_determinist(l, &mut rng, d)
+        Self::new_determinist(l, &mut rng, d)
     }
 
     /// Create a new cold configuration for the electrical field, i.e. all E ar set to 0.
@@ -1133,6 +1277,8 @@ impl<const D: usize> EField<D> {
     /// # Ok(())
     /// # }
     /// ```
+    #[inline]
+    #[must_use]
     pub fn new_cold(l: &LatticeCyclic<D>) -> Self {
         let p1 = Su3Adjoint::new_from_array([0_f64; 8]);
         Self {
@@ -1141,6 +1287,8 @@ impl<const D: usize> EField<D> {
     }
 
     /// Get `E(point) = [E_x(point), E_y(point), E_z(point)]`.
+    #[inline]
+    #[must_use]
     pub fn e_vec(
         &self,
         point: &LatticePoint<D>,
@@ -1151,6 +1299,8 @@ impl<const D: usize> EField<D> {
 
     /// Get `E_{dir}(point)`. The sign of the direction does not change the output. i.e.
     /// `E_{-dir}(point) = E_{dir}(point)`.
+    #[inline]
+    #[must_use]
     pub fn e_field(
         &self,
         point: &LatticePoint<D>,
@@ -1158,19 +1308,19 @@ impl<const D: usize> EField<D> {
         l: &LatticeCyclic<D>,
     ) -> Option<&Su3Adjoint> {
         let value = self.e_vec(point, l);
-        match value {
-            Some(vec) => vec.get(dir.index()),
-            None => None,
-        }
+        value.and_then(|vec| vec.get(dir.index()))
     }
 
     /// Returns wether there is no data
+    #[inline]
+    #[must_use]
     pub fn is_empty(&self) -> bool {
         self.data.is_empty()
     }
 
     /// Return the Gauss parameter `G(x) = \sum_i E_i(x) - U_{-i}(x) E_i(x - i) U^\dagger_{-i}(x)`.
     #[inline]
+    #[must_use]
     pub fn gauss(
         &self,
         link_matrix: &LinkMatrix,
@@ -1196,6 +1346,7 @@ impl<const D: usize> EField<D> {
 
     /// Get the deviation from the Gauss law
     #[inline]
+    #[must_use]
     pub fn gauss_sum_div(
         &self,
         link_matrix: &LinkMatrix,
@@ -1261,7 +1412,9 @@ impl<const D: usize> EField<D> {
     /// # }
     /// ```
     #[allow(clippy::as_conversions)] // no try into for f64
+    #[allow(clippy::cast_precision_loss)]
     #[inline]
+    #[must_use]
     pub fn project_to_gauss(
         &self,
         link_matrix: &LinkMatrix,
@@ -1297,7 +1450,7 @@ impl<const D: usize> EField<D> {
     ///
     /// # Panic
     /// panics if the link matrix and lattice is not of the correct size.
-    #[inline]
+    #[inline] // REVIEW
     fn project_to_gauss_step(&self, link_matrix: &LinkMatrix, lattice: &LatticeCyclic<D>) -> Self {
         /// see <https://arxiv.org/pdf/1512.02374.pdf>
         // TODO verify
@@ -1307,20 +1460,22 @@ impl<const D: usize> EField<D> {
             .collect::<Vec<LatticePoint<D>>>()
             .par_iter()
             .map(|point| {
-                let e = self.e_vec(point, lattice).unwrap();
+                let e = self.e_vec(point, lattice).expect("e vec not founc");
                 SVector::<_, D>::from_fn(|index_dir, _| {
                     let dir = Direction::<D>::positive_directions()[index_dir];
                     let u = link_matrix
                         .matrix(&LatticeLink::new(*point, dir), lattice)
-                        .unwrap();
-                    let gauss = self.gauss(link_matrix, point, lattice).unwrap();
+                        .expect("matrix not found");
+                    let gauss = self
+                        .gauss(link_matrix, point, lattice)
+                        .expect("gauss not found");
                     let gauss_p = self
                         .gauss(
                             link_matrix,
                             &lattice.add_point_direction(*point, &dir),
                             lattice,
                         )
-                        .unwrap();
+                        .expect("gauss not found");
                     Su3Adjoint::new(Vector8::from_fn(|index, _| {
                         2_f64
                             * (su3::GENERATORS[index]
@@ -1338,24 +1493,28 @@ impl<const D: usize> EField<D> {
 }
 
 impl<const D: usize> AsRef<Vec<SVector<Su3Adjoint, D>>> for EField<D> {
+    #[inline]
     fn as_ref(&self) -> &Vec<SVector<Su3Adjoint, D>> {
         self.as_vec()
     }
 }
 
 impl<const D: usize> AsMut<Vec<SVector<Su3Adjoint, D>>> for EField<D> {
+    #[inline]
     fn as_mut(&mut self) -> &mut Vec<SVector<Su3Adjoint, D>> {
         self.data_mut()
     }
 }
 
 impl<const D: usize> AsRef<[SVector<Su3Adjoint, D>]> for EField<D> {
+    #[inline]
     fn as_ref(&self) -> &[SVector<Su3Adjoint, D>] {
         self.as_slice()
     }
 }
 
 impl<const D: usize> AsMut<[SVector<Su3Adjoint, D>]> for EField<D> {
+    #[inline]
     fn as_mut(&mut self) -> &mut [SVector<Su3Adjoint, D>] {
         self.as_slice_mut()
     }
@@ -1365,6 +1524,7 @@ impl<'a, const D: usize> IntoIterator for &'a EField<D> {
     type IntoIter = <&'a Vec<SVector<Su3Adjoint, D>> as IntoIterator>::IntoIter;
     type Item = &'a SVector<Su3Adjoint, D>;
 
+    #[inline]
     fn into_iter(self) -> Self::IntoIter {
         self.data.iter()
     }
@@ -1374,6 +1534,7 @@ impl<'a, const D: usize> IntoIterator for &'a mut EField<D> {
     type IntoIter = <&'a mut Vec<SVector<Su3Adjoint, D>> as IntoIterator>::IntoIter;
     type Item = &'a mut SVector<Su3Adjoint, D>;
 
+    #[inline]
     fn into_iter(self) -> Self::IntoIter {
         self.data.iter_mut()
     }
@@ -1382,12 +1543,14 @@ impl<'a, const D: usize> IntoIterator for &'a mut EField<D> {
 impl<const D: usize> Index<usize> for EField<D> {
     type Output = SVector<Su3Adjoint, D>;
 
+    #[inline]
     fn index(&self, pos: usize) -> &Self::Output {
         &self.data[pos]
     }
 }
 
 impl<const D: usize> IndexMut<usize> for EField<D> {
+    #[inline]
     fn index_mut(&mut self, pos: usize) -> &mut Self::Output {
         &mut self.data[pos]
     }
@@ -1397,6 +1560,7 @@ impl<A, const D: usize> FromIterator<A> for EField<D>
 where
     Vec<SVector<Su3Adjoint, D>>: FromIterator<A>,
 {
+    #[inline]
     fn from_iter<T>(iter: T) -> Self
     where
         T: IntoIterator<Item = A>,
@@ -1410,6 +1574,7 @@ where
     Vec<SVector<Su3Adjoint, D>>: FromParallelIterator<A>,
     A: Send,
 {
+    #[inline]
     fn from_par_iter<I>(par_iter: I) -> Self
     where
         I: IntoParallelIterator<Item = A>,
@@ -1423,6 +1588,7 @@ where
     Vec<SVector<Su3Adjoint, D>>: ParallelExtend<T>,
     T: Send,
 {
+    #[inline]
     fn par_extend<I>(&mut self, par_iter: I)
     where
         I: IntoParallelIterator<Item = T>,
@@ -1435,6 +1601,7 @@ impl<A, const D: usize> Extend<A> for EField<D>
 where
     Vec<SVector<Su3Adjoint, D>>: Extend<A>,
 {
+    #[inline]
     fn extend<T>(&mut self, iter: T)
     where
         T: IntoIterator<Item = A>,
@@ -1445,20 +1612,26 @@ where
 
 #[cfg(test)]
 mod test {
+    use std::error::Error;
+
     use approx::*;
+    use rand::rngs::StdRng;
     use rand::SeedableRng;
+    use rand_distr::Uniform;
 
     use super::super::{lattice::*, Complex};
     use super::*;
+    use crate::error::{ImplementationError, LatticeInitializationError};
+    use crate::su3::su3_exp_r;
 
     const EPSILON: f64 = 0.000_000_001_f64;
     const SEED_RNG: u64 = 0x45_78_93_f4_4a_b0_67_f0;
 
     #[test]
-    fn test_get_e_field_pos_neg() {
+    fn test_get_e_field_pos_neg() -> Result<(), LatticeInitializationError> {
         use super::super::lattice;
 
-        let l = LatticeCyclic::new(1_f64, 4).unwrap();
+        let l = LatticeCyclic::new(1_f64, 4)?;
         let e = EField::new(vec![SVector::<_, 4>::from([
             Su3Adjoint::from([1_f64; 8]),
             Su3Adjoint::from([2_f64; 8]),
@@ -1477,14 +1650,15 @@ mod test {
                 &l
             )
         );
+        Ok(())
     }
 
     #[test]
     #[allow(clippy::eq_op)]
     #[allow(clippy::op_ref)]
     fn test_su3_adj() {
-        let mut rng = rand::rngs::StdRng::seed_from_u64(SEED_RNG);
-        let d = rand::distributions::Uniform::from(-1_f64..1_f64);
+        let mut rng = StdRng::seed_from_u64(SEED_RNG);
+        let d = Uniform::from(-1_f64..1_f64);
         for _ in 0_u32..100_u32 {
             let v = Su3Adjoint::random(&mut rng, &d);
             let m = v.to_matrix();
@@ -1526,9 +1700,8 @@ mod test {
             assert_eq!(adj_2, -adj_4);
         }
 
-        use crate::su3::su3_exp_r;
-        let mut rng = rand::rngs::StdRng::seed_from_u64(SEED_RNG);
-        let d = rand::distributions::Uniform::from(-1_f64..1_f64);
+        let mut rng = StdRng::seed_from_u64(SEED_RNG);
+        let d = Uniform::from(-1_f64..1_f64);
         for _ in 0_u32..10_u32 {
             let v = Su3Adjoint::random(&mut rng, &d);
             assert_eq!(su3_exp_r(v), v.exp());
@@ -1536,15 +1709,15 @@ mod test {
     }
 
     #[test]
-    fn link_matrix() {
-        let lattice = LatticeCyclic::<3>::new(1_f64, 4).unwrap();
+    fn link_matrix() -> Result<(), Box<dyn Error>> {
+        let lattice = LatticeCyclic::<3>::new(1_f64, 4)?;
         match LinkMatrix::new_random_threaded(&lattice, 0) {
             Err(ThreadError::ThreadNumberIncorrect) => {}
             _ => panic!("unexpected output"),
         }
         let link_s = LinkMatrix::new_random_threaded(&lattice, 2);
         assert!(link_s.is_ok());
-        let mut link = link_s.unwrap();
+        let mut link = link_s?;
         assert!(!link.is_empty());
         let l2 = LinkMatrix::new(vec![]);
         assert!(l2.is_empty());
@@ -1553,50 +1726,63 @@ mod test {
         let _: &Vec<_> = link.as_ref();
         let _: &mut [_] = link.as_mut();
         let _: &mut Vec<_> = link.as_mut();
-        let _ = link.iter();
-        let _ = link.iter_mut();
-        let _ = (&link).into_iter();
-        let _ = (&mut link).into_iter();
+        #[allow(clippy::let_underscore_must_use)]
+        #[allow(clippy::let_underscore_untyped)]
+        {
+            let _ = link.iter();
+            let _ = link.iter_mut();
+            let _ = (&link).into_iter();
+            let _ = (&mut link).into_iter();
+        }
+        Ok(())
     }
 
     #[test]
-    fn e_field() {
-        let lattice = LatticeCyclic::<3>::new(1_f64, 4).unwrap();
+    fn e_field() -> Result<(), Box<dyn Error>> {
+        let lattice = LatticeCyclic::<3>::new(1_f64, 4)?;
         let e_field_s = LinkMatrix::new_random_threaded(&lattice, 2);
         assert!(e_field_s.is_ok());
-        let mut e_field = e_field_s.unwrap();
+        let mut e_field = e_field_s?;
 
         let _: &[_] = e_field.as_ref();
         let _: &Vec<_> = e_field.as_ref();
         let _: &mut [_] = e_field.as_mut();
         let _: &mut Vec<_> = e_field.as_mut();
-        let _ = e_field.iter();
-        let _ = e_field.iter_mut();
-        let _ = (&e_field).into_iter();
-        let _ = (&mut e_field).into_iter();
+        #[allow(clippy::let_underscore_must_use)]
+        #[allow(clippy::let_underscore_untyped)]
+        {
+            let _ = e_field.iter();
+            let _ = e_field.iter_mut();
+            let _ = (&e_field).into_iter();
+            let _ = (&mut e_field).into_iter();
+        }
+        Ok(())
     }
 
+    #[allow(clippy::too_many_lines)]
     #[test]
-    fn magnetic_field() {
-        let lattice = LatticeCyclic::<3>::new(1_f64, 4).unwrap();
+    fn magnetic_field() -> Result<(), Box<dyn Error>> {
+        let lattice = LatticeCyclic::<3>::new(1_f64, 4)?;
         let mut link_matrix = LinkMatrix::new_cold(&lattice);
         let point = LatticePoint::from([0, 0, 0]);
-        let dir_x = Direction::new(0, true).unwrap();
-        let dir_y = Direction::new(1, true).unwrap();
-        let dir_z = Direction::new(2, true).unwrap();
+        let dir_x = Direction::new(0, true).ok_or(ImplementationError::OptionWithUnexpectedNone)?;
+        let dir_y = Direction::new(1, true).ok_or(ImplementationError::OptionWithUnexpectedNone)?;
+        let dir_z = Direction::new(2, true).ok_or(ImplementationError::OptionWithUnexpectedNone)?;
         let clover = link_matrix
             .clover(&point, &dir_x, &dir_y, &lattice)
-            .unwrap();
+            .ok_or(ImplementationError::OptionWithUnexpectedNone)?;
         assert_eq_matrix!(CMatrix3::identity() * Complex::from(4_f64), clover, EPSILON);
         let f = link_matrix
             .f_mu_nu(&point, &dir_x, &dir_y, &lattice)
-            .unwrap();
+            .ok_or(ImplementationError::OptionWithUnexpectedNone)?;
         assert_eq_matrix!(CMatrix3::zeros(), f, EPSILON);
         let b = link_matrix
             .magnetic_field(&point, &dir_x, &lattice)
-            .unwrap();
+            .ok_or(ImplementationError::OptionWithUnexpectedNone)?;
         assert_eq_matrix!(CMatrix3::zeros(), b, EPSILON);
-        let b_vec = link_matrix.magnetic_field_vec(&point, &lattice).unwrap();
+        let b_vec = link_matrix
+            .magnetic_field_vec(&point, &lattice)
+            .ok_or(ImplementationError::OptionWithUnexpectedNone)?;
         for i in &b_vec {
             assert_eq_matrix!(CMatrix3::zeros(), i, EPSILON);
         }
@@ -1604,7 +1790,7 @@ mod test {
         link_matrix[0] = CMatrix3::identity() * Complex::new(0_f64, 1_f64);
         let clover = link_matrix
             .clover(&point, &dir_x, &dir_y, &lattice)
-            .unwrap();
+            .ok_or(ImplementationError::OptionWithUnexpectedNone)?;
         assert_eq_matrix!(
             CMatrix3::identity() * Complex::new(2_f64, 0_f64),
             clover,
@@ -1612,7 +1798,7 @@ mod test {
         );
         let clover = link_matrix
             .clover(&point, &dir_y, &dir_x, &lattice)
-            .unwrap();
+            .ok_or(ImplementationError::OptionWithUnexpectedNone)?;
         assert_eq_matrix!(
             CMatrix3::identity() * Complex::new(2_f64, 0_f64),
             clover,
@@ -1620,7 +1806,7 @@ mod test {
         );
         let f = link_matrix
             .f_mu_nu(&point, &dir_x, &dir_y, &lattice)
-            .unwrap();
+            .ok_or(ImplementationError::OptionWithUnexpectedNone)?;
         assert_eq_matrix!(
             CMatrix3::identity() * Complex::new(0_f64, 0_f64),
             f,
@@ -1628,26 +1814,29 @@ mod test {
         );
         let b = link_matrix
             .magnetic_field(&point, &dir_x, &lattice)
-            .unwrap();
+            .ok_or(ImplementationError::OptionWithUnexpectedNone)?;
         assert_eq_matrix!(CMatrix3::zeros(), b, EPSILON);
-        let b_vec = link_matrix.magnetic_field_vec(&point, &lattice).unwrap();
+        let b_vec = link_matrix
+            .magnetic_field_vec(&point, &lattice)
+            .ok_or(ImplementationError::OptionWithUnexpectedNone)?;
         for i in &b_vec {
             assert_eq_matrix!(CMatrix3::zeros(), i, EPSILON);
         }
         assert_eq_matrix!(
             link_matrix
                 .magnetic_field_link(&LatticeLink::new(point, dir_x), &lattice)
-                .unwrap(),
+                .ok_or(ImplementationError::OptionWithUnexpectedNone)?,
             b,
             EPSILON
         );
         //--
         let mut link_matrix = LinkMatrix::new_cold(&lattice);
-        let link = LatticeLinkCanonical::new([1, 0, 0].into(), dir_y).unwrap();
+        let link = LatticeLinkCanonical::new([1, 0, 0].into(), dir_y)
+            .ok_or(ImplementationError::OptionWithUnexpectedNone)?;
         link_matrix[link.to_index(&lattice)] = CMatrix3::identity() * Complex::new(0_f64, 1_f64);
         let clover = link_matrix
             .clover(&point, &dir_x, &dir_y, &lattice)
-            .unwrap();
+            .ok_or(ImplementationError::OptionWithUnexpectedNone)?;
         assert_eq_matrix!(
             CMatrix3::identity() * Complex::new(3_f64, 1_f64),
             clover,
@@ -1655,7 +1844,7 @@ mod test {
         );
         let clover = link_matrix
             .clover(&point, &dir_y, &dir_x, &lattice)
-            .unwrap();
+            .ok_or(ImplementationError::OptionWithUnexpectedNone)?;
         assert_eq_matrix!(
             CMatrix3::identity() * Complex::new(3_f64, -1_f64),
             clover,
@@ -1663,7 +1852,7 @@ mod test {
         );
         let f = link_matrix
             .f_mu_nu(&point, &dir_x, &dir_y, &lattice)
-            .unwrap();
+            .ok_or(ImplementationError::OptionWithUnexpectedNone)?;
         assert_eq_matrix!(
             CMatrix3::identity() * Complex::new(0_f64, 0.25_f64),
             f,
@@ -1671,22 +1860,22 @@ mod test {
         );
         let b = link_matrix
             .magnetic_field(&point, &dir_x, &lattice)
-            .unwrap();
+            .ok_or(ImplementationError::OptionWithUnexpectedNone)?;
         assert_eq_matrix!(CMatrix3::zeros(), b, EPSILON);
         assert_eq_matrix!(
             link_matrix
                 .magnetic_field_link(&LatticeLink::new(point, dir_x), &lattice)
-                .unwrap(),
+                .ok_or(ImplementationError::OptionWithUnexpectedNone)?,
             b,
             EPSILON
         );
         let b = link_matrix
             .magnetic_field(&point, &dir_z, &lattice)
-            .unwrap();
+            .ok_or(ImplementationError::OptionWithUnexpectedNone)?;
         assert_eq_matrix!(
             link_matrix
                 .magnetic_field_link(&LatticeLink::new(point, dir_z), &lattice)
-                .unwrap(),
+                .ok_or(ImplementationError::OptionWithUnexpectedNone)?,
             b,
             EPSILON
         );
@@ -1697,17 +1886,20 @@ mod test {
         );
         let b_2 = link_matrix
             .magnetic_field(&[4, 0, 0].into(), &dir_z, &lattice)
-            .unwrap();
+            .ok_or(ImplementationError::OptionWithUnexpectedNone)?;
         assert_eq_matrix!(b, b_2, EPSILON);
-        let b_vec = link_matrix.magnetic_field_vec(&point, &lattice).unwrap();
+        let b_vec = link_matrix
+            .magnetic_field_vec(&point, &lattice)
+            .ok_or(ImplementationError::OptionWithUnexpectedNone)?;
         for (index, m) in b_vec.iter().enumerate() {
             if index == 2 {
                 assert_eq_matrix!(m, b, EPSILON);
-            }
-            else {
+            } else {
                 assert_eq_matrix!(CMatrix3::zeros(), m, EPSILON);
             }
         }
+
+        Ok(())
     }
 
     #[test]
