@@ -4,9 +4,10 @@
 
 use std::cmp::Ordering;
 use std::convert::TryInto;
+use std::fmt::{self, Display};
 use std::ops::{Mul, MulAssign, Neg};
 
-use approx::*;
+use approx::abs_diff_eq;
 #[cfg(feature = "serde-serialize")]
 use serde::{Deserialize, Serialize};
 
@@ -35,11 +36,12 @@ pub const MAX_NUMBER_FACTORIAL: usize = 34;
 /// assert!(!overflowed);
 /// ```
 #[allow(clippy::as_conversions)] // constant function cant use try into
+#[inline]
+#[must_use]
 pub const fn factorial(n: usize) -> FactorialNumber {
     if n == 0 {
         1
-    }
-    else {
+    } else {
         n as FactorialNumber * factorial(n - 1)
     }
 }
@@ -53,6 +55,7 @@ pub struct FactorialStorageDyn {
 }
 
 impl Default for FactorialStorageDyn {
+    #[inline]
     fn default() -> Self {
         Self::new()
     }
@@ -60,6 +63,8 @@ impl Default for FactorialStorageDyn {
 
 impl FactorialStorageDyn {
     /// Create a new object with an empty storage.
+    #[inline]
+    #[must_use]
     pub const fn new() -> Self {
         Self { data: Vec::new() }
     }
@@ -73,8 +78,10 @@ impl FactorialStorageDyn {
     /// f.build_storage(6);
     /// assert_eq!(*f.try_factorial(6).unwrap(), 720);
     /// ```
+    #[allow(clippy::let_underscore_must_use)]
+    #[inline]
     pub fn build_storage(&mut self, value: usize) {
-        self.factorial(value);
+        let _: FactorialNumber = self.factorial(value);
     }
 
     /// Get the factorial number. If it is not already computed build internal storage
@@ -89,6 +96,9 @@ impl FactorialStorageDyn {
     /// assert_eq!(f.factorial(6), 720);
     /// assert_eq!(f.factorial(4), 24);
     /// ```
+    #[allow(clippy::missing_panics_doc)] // never panics
+    #[inline]
+    #[must_use]
     pub fn factorial(&mut self, value: usize) -> FactorialNumber {
         let mut len = self.data.len();
         if len == 0 {
@@ -98,9 +108,11 @@ impl FactorialStorageDyn {
         if len > value {
             return self.data[value];
         }
-        for i in len..value + 1 {
-            self.data
-                .push(self.data[i - 1] * TryInto::<FactorialNumber>::try_into(i).unwrap());
+        for i in len..=value {
+            self.data.push(
+                self.data[i - 1]
+                    * TryInto::<FactorialNumber>::try_into(i).expect("conversion always possible"),
+            );
         }
         self.data[value]
     }
@@ -115,6 +127,8 @@ impl FactorialStorageDyn {
     /// assert_eq!(*f.try_factorial(4).unwrap(), 24);
     /// assert_eq!(f.try_factorial(6), None);
     /// ```
+    #[inline]
+    #[must_use]
     pub fn try_factorial(&self, value: usize) -> Option<&FactorialNumber> {
         self.data.get(value)
     }
@@ -128,34 +142,42 @@ impl FactorialStorageDyn {
     /// assert_eq!(f.factorial_no_storage(6), 720);
     /// assert_eq!(f.try_factorial(6), None);
     /// ```
+    #[allow(clippy::missing_panics_doc)] // never panics
+    #[inline]
+    #[must_use]
     pub fn factorial_no_storage(&self, value: usize) -> FactorialNumber {
-        let mut value_m: FactorialNumber = self.data[value.min(self.data.len() - 1)];
+        let mut value_m = self.data[value.min(self.data.len() - 1)];
         for i in self.data.len() - 1..value {
-            value_m *= TryInto::<FactorialNumber>::try_into(i + 1).unwrap();
+            value_m *=
+                TryInto::<FactorialNumber>::try_into(i + 1).expect("conversion always possible");
         }
         value_m
     }
 }
 
 /// Represent a sign.
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Copy)]
+#[derive(Clone, Debug, PartialEq, Eq, Hash, Copy, Default)]
 #[cfg_attr(feature = "serde-serialize", derive(Serialize, Deserialize))]
+#[allow(clippy::exhaustive_enums)]
 pub enum Sign {
     /// Strictly negative number (non zero)
     Negative,
     /// Strictly positive number ( non zero)
     Positive,
     /// Zero (or very close to zero)
+    #[default]
     Zero,
 }
 
 impl Sign {
     /// return a f64 form the sign `(-1_f64, 0_f64, 1_f64)`.
+    #[inline]
+    #[must_use]
     pub const fn to_f64(self) -> f64 {
         match self {
-            Sign::Negative => -1_f64,
-            Sign::Positive => 1_f64,
-            Sign::Zero => 0_f64,
+            Self::Negative => -1_f64,
+            Self::Positive => 1_f64,
+            Self::Zero => 0_f64,
         }
     }
 
@@ -163,64 +185,59 @@ impl Sign {
     ///
     /// If the value is very close to zero but not quite the sing will nonetheless be [`Sign::Zero`].
     /// If f is NaN the sing will be [`Sign::Zero`].
+    #[inline]
+    #[must_use]
     pub fn sign_f64(f: f64) -> Self {
         if abs_diff_eq!(f, 0_f64) || f.is_nan() {
-            Sign::Zero
-        }
-        else if f > 0_f64 {
-            Sign::Positive
-        }
-        else {
-            Sign::Negative
+            Self::Zero
+        } else if f > 0_f64 {
+            Self::Positive
+        } else {
+            Self::Negative
         }
     }
 
     /// Convert the sign to an i8.
+    #[inline]
+    #[must_use]
     pub const fn to_i8(self) -> i8 {
         match self {
-            Sign::Negative => -1_i8,
-            Sign::Positive => 1_i8,
-            Sign::Zero => 0_i8,
+            Self::Negative => -1_i8,
+            Self::Positive => 1_i8,
+            Self::Zero => 0_i8,
         }
     }
 
     /// Get the sign of the given [`i8`]
-    #[allow(clippy::comparison_chain)] // Cannot use cmp in const function
+    #[inline]
+    #[must_use]
     pub const fn sign_i8(n: i8) -> Self {
         if n == 0 {
-            Sign::Zero
-        }
-        else if n > 0 {
-            Sign::Positive
-        }
-        else {
-            Sign::Negative
+            Self::Zero
+        } else if n > 0 {
+            Self::Positive
+        } else {
+            Self::Negative
         }
     }
 
     /// Returns the sign of `a - b`, where `a` and `b` are usize
-    #[allow(clippy::comparison_chain)]
+    #[inline]
+    #[must_use]
     pub const fn sign_from_diff(a: usize, b: usize) -> Self {
         if a == b {
-            Sign::Zero
-        }
-        else if a > b {
-            Sign::Positive
-        }
-        else {
-            Sign::Negative
+            Self::Zero
+        } else if a > b {
+            Self::Positive
+        } else {
+            Self::Negative
         }
     }
 }
 
-impl Default for Sign {
-    fn default() -> Self {
-        Sign::Zero
-    }
-}
-
-impl std::fmt::Display for Sign {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+impl Display for Sign {
+    #[inline]
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
             Self::Positive => write!(f, "positive"),
             Self::Zero => write!(f, "zero"),
@@ -230,37 +247,42 @@ impl std::fmt::Display for Sign {
 }
 
 impl From<Sign> for f64 {
-    fn from(s: Sign) -> f64 {
+    #[inline]
+    fn from(s: Sign) -> Self {
         s.to_f64()
     }
 }
 
 impl From<f64> for Sign {
-    fn from(f: f64) -> Sign {
-        Sign::sign_f64(f)
+    #[inline]
+    fn from(f: f64) -> Self {
+        Self::sign_f64(f)
     }
 }
 
 impl From<Sign> for i8 {
-    fn from(s: Sign) -> i8 {
+    #[inline]
+    fn from(s: Sign) -> Self {
         s.to_i8()
     }
 }
 
 impl From<i8> for Sign {
-    fn from(i: i8) -> Sign {
-        Sign::sign_i8(i)
+    #[inline]
+    fn from(i: i8) -> Self {
+        Self::sign_i8(i)
     }
 }
 
 impl Neg for Sign {
     type Output = Self;
 
+    #[inline]
     fn neg(self) -> Self::Output {
         match self {
-            Sign::Positive => Sign::Negative,
-            Sign::Zero => Sign::Zero,
-            Sign::Negative => Sign::Positive,
+            Self::Positive => Self::Negative,
+            Self::Zero => Self::Zero,
+            Self::Negative => Self::Positive,
         }
     }
 }
@@ -268,28 +290,32 @@ impl Neg for Sign {
 impl Mul for Sign {
     type Output = Self;
 
+    #[inline]
     fn mul(self, rhs: Self) -> Self {
         match (self, rhs) {
-            (Sign::Negative, Sign::Negative) | (Sign::Positive, Sign::Positive) => Sign::Positive,
-            (Sign::Zero, _) | (_, Sign::Zero) => Sign::Zero,
-            (Sign::Positive, Sign::Negative) | (Sign::Negative, Sign::Positive) => Sign::Negative,
+            (Self::Negative, Self::Negative) | (Self::Positive, Self::Positive) => Self::Positive,
+            (Self::Zero, _) | (_, Self::Zero) => Self::Zero,
+            (Self::Positive, Self::Negative) | (Self::Negative, Self::Positive) => Self::Negative,
         }
     }
 }
 
-impl MulAssign<Sign> for Sign {
+impl MulAssign<Self> for Sign {
+    #[inline]
     fn mul_assign(&mut self, rhs: Self) {
         *self = *self * rhs;
     }
 }
 
 impl PartialOrd for Sign {
+    #[inline]
     fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
         Some(self.cmp(other))
     }
 }
 
 impl Ord for Sign {
+    #[inline]
     fn cmp(&self, other: &Self) -> Ordering {
         self.to_i8().cmp(&other.to_i8())
     }
@@ -303,6 +329,8 @@ impl Ord for Sign {
 /// assert_eq!(Sign::Negative, levi_civita(&[2, 1, 3]));
 /// assert_eq!(Sign::Zero, levi_civita(&[2, 2, 3]));
 /// ```
+#[inline]
+#[must_use]
 pub const fn levi_civita(index: &[usize]) -> Sign {
     let mut prod = 1_i8;
     let mut i = 0_usize;
@@ -321,25 +349,27 @@ pub const fn levi_civita(index: &[usize]) -> Sign {
 mod test {
     use super::*;
 
+    /// test that the factorial pass for [`MAX_NUMBER_FACTORIAL`]
     #[allow(clippy::missing_const_for_fn)]
+    #[allow(clippy::let_underscore_must_use)]
     #[test]
-    /// test that the factorial pass for MAX_NUMBER_FACTORIAL
     fn test_factorial_pass() {
-        factorial(MAX_NUMBER_FACTORIAL);
+        let _: FactorialNumber = factorial(MAX_NUMBER_FACTORIAL);
     }
 
     #[allow(clippy::missing_const_for_fn)]
+    #[allow(clippy::let_underscore_must_use)]
     #[test]
     #[should_panic]
-    #[cfg(not(feature = "no-overflow-test"))]
-    /// test that the factorial overflow for MAX_NUMBER_FACTORIAL + 1
+    #[cfg(feature = "overflow-test")]
+    /// test that the factorial overflow for `MAX_NUMBER_FACTORIAL + 1`
     fn test_factorial_bigger() {
-        factorial(MAX_NUMBER_FACTORIAL + 1);
+        let _: FactorialNumber = factorial(MAX_NUMBER_FACTORIAL + 1);
     }
 
     #[test]
     #[should_panic]
-    /// test that the factorial overflow for MAX_NUMBER_FACTORIAL + 1
+    /// test that the factorial overflow for `MAX_NUMBER_FACTORIAL + 1`
     fn test_factorial_overflow() {
         let n = factorial(MAX_NUMBER_FACTORIAL);
         let (_, overflowed) = n.overflowing_mul(MAX_NUMBER_FACTORIAL as u128 + 1);
@@ -385,6 +415,7 @@ mod test {
     }
 
     #[allow(clippy::cognitive_complexity)]
+    #[allow(clippy::float_cmp)]
     #[test]
     fn sign() {
         assert_eq!(Sign::sign_f64(0_f64).to_f64(), 0_f64);
